+++ src/clibard.nim
@@ -9,18 +9,21 @@ from std/random import randomize, rand
 import pkg/bard
 import pkg/gookie
 
-proc startNewBardChat: BardAiChat =
-  echo "Trying to get your Google session"
+proc startNewBardChat(silent = false): BardAiChat =
+  if not silent:
+    echo "Trying to get your Google session"
   let cookies = parseCookies getGoogleCookies()
   if not (cookies.hasKey("__Secure-1PSID") and cookies.hasKey "__Secure-1PSIDTS"):
     quit "Cannot get session. Please login into Google account in browser"
-  echo "Session successfully got!"
-  echo "Creating new Google Bard instance"
+  if not silent:
+    echo "Session successfully got!"
+    echo "Creating new Google Bard instance"
   let ai = waitFor newBardAi(
     psid = cookies["__Secure-1PSID"],
     psidts = cookies["__Secure-1PSIDTS"]
   )
-  echo "Google Bard instance successfully created!"
+  if not silent:
+    echo "Google Bard instance successfully created!\n"
   result = newBardAiChat ai
 
 
@@ -34,34 +37,32 @@ proc typingEcho(s: string; instant = false; fast = false) =
       sleep rand(
         if not fast:
           case ch:
-          of '\n': 10..50
-          of ' ': 3..10
-          else: 5..20
+          of '\n': 60..100
+          of ' ': 20..60
+          else: 10..30
         else: 1..5
       )
 
-proc cliPrompt(texts: seq[string]; instant = false; fast = false) =
+proc cliPrompt(texts: seq[string]; instant = true; fast = false; silent = true) =
   ## Prompts to Google Bard
-  var chat = startNewBardChat()
+  var chat = startNewBardChat silent
   let text = texts.join " "
   try:
     let response = waitFor chat.prompt text
-    echo ""
     typingEcho response.text, instant
   except BardExpiredSession:
-    echo getCurrentExceptionMsg()
-    cliPrompt(@[text])
+    cliPrompt(@[text], instant, fast, silent)
 
-proc cliChat(instant = false; fast = false) =
+proc cliChat(instant = false; fast = false; silent = false) =
   ## Start chat with Google Bard
   ## 
-  ## Close with ".exit"
-  var chat = startNewBardChat()
-  echo "\l==Chat started==\l"
+  ## Close with "exit"
+  var chat = startNewBardChat silent
+  echo "==Chat started==\l"
   while true:
     stdout.write "You: "
     let text = readLine stdin
-    if text == ".exit":
+    if text == "exit":
       break
     try:
       let response = waitFor chat.prompt text
@@ -69,7 +70,7 @@ proc cliChat(instant = false; fast = false) =
       typingEcho response.text, instant
       echo "\l"
     except BardExpiredSession:
-      cliPrompt(@[text])
+      cliPrompt(@[text], false, fast, silent)
 
 when isMainModule:
   import pkg/cligen